requirements.txt
google-generativeai
python-dotenv
yfinance
pandas
requests
streamlit